refactor: 참여 코드 인증 관련 로직 훅 분리
